# Conflicts:
#	MIDITrail/DXPrimitive.cpp
#	MIDITrail/MIDITrail.rc
#	MIDITrail/MIDITrailApp.cpp
#	MIDITrail/MIDITrailApp.h
#	MIDITrail/MTFont2Bmp.cpp
#	MIDITrail/MTNoteDesign.h
#	MIDITrail/MTPianoKeyboard.cpp
#	MIDITrail/MTPianoKeyboard.h
#	MIDITrail/MTPianoKeyboardCtrl.h
#	MIDITrail/MTPianoKeyboardCtrlMod.cpp
#	MIDITrail/MTPianoKeyboardCtrlMod.h
#	MIDITrail/MTPianoKeyboardDesign.h
#	MIDITrail/MTPianoKeyboardDesignMod.cpp
#	MIDITrail/MTPianoKeyboardDesignMod.h
#	MIDITrail/MTPianoKeyboardMod.cpp
#	MIDITrail/MTPianoKeyboardMod.h
#	MIDITrail/MTScenePianoRoll3D.h
#	MIDITrail/MTScenePianoRoll3DMod.cpp
#	MIDITrail/MTScenePianoRoll3DMod.h
#	MIDITrail/resource.h

+++ MIDITrail/MTDashboardLive.cpp
@@ -4,7 +4,7 @@
 //
 // ライブモニタ用ダッシュボード描画クラス
 //
-// Copyright (C) 2012 WADA Masashi. All Rights Reserved.
+// Copyright (C) 2012-2014 WADA Masashi. All Rights Reserved.
 //
 //******************************************************************************
 
@@ -177,7 +177,7 @@ int MTDashboardLive::_GetCounterPos(
 	
 	//文字サイズ
 	charHeight = th;
-	charWidth = tw / _tcslen(MTDASHBOARDLIVE_COUNTER_CHARS);
+	charWidth = tw / (unsigned long)_tcslen(MTDASHBOARDLIVE_COUNTER_CHARS);
 	
 	//拡大率1.0のキャプションサイズ
 	captionWidth = (unsigned long)(charWidth * MTDASHBOARDLIVE_COUNTER_SIZE);
